+++ src/main/java/uk/co/transactis/Shape.java
@@ -1,5 +1,5 @@
 package uk.co.transactis;
 
 public interface Shape {
-    public float getArea();
+    public double getArea();
 }
