+++ src/main/java/uk/co/transactis/Square.java
@@ -3,7 +3,7 @@ package uk.co.transactis;
 
 public class Square implements Shape {
     
-    public float getArea() {
+    public double getArea() {
     	return 0;
     }
 }
